Create multiple packages for Java Classes
