--- data/tasks.txt
@@ -1,3 +1,0 @@
-<<<<<<< HEAD
-=======
->>>>>>> branch-Level-8
