--- src/main/java/Command.java
@@ -1,7 +1,0 @@
-import java.io.IOException;
-
-abstract class Command {
-    abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException, IOException;
-
-    abstract boolean isExit();
-}
