+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,12 @@
+package duke.command;
+
+import duke.util.Storage;
+import duke.util.Ui;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.TaskList;
+import duke.task.Todo;
+
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
