+++ src/main/java/duke/command/MarkCommand.java
@@ -1,5 +1,11 @@
-import java.io.IOException;
+package duke.command;
 
+import duke.util.Storage;
+import duke.util.Ui;
+import duke.exception.DukeException;
+import duke.task.TaskList;
+
+import java.io.IOException;
 public class MarkCommand extends Command {
     private int taskNumber;
